requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
urllib3==1.26.15
